+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.1
+
+- Populate the calculated-value `Tensor.length`.
+- Calculates the layour block count using standard tensor names.
+
 ## 0.1.0
 
 - Initial version.